Renamed custom button and text span files to avoid confusing with custard buttons/text spans
